+++ lib/adhearsion-asr/controller_methods.rb
@@ -42,8 +42,9 @@ module AdhearsionASR
 
       options[:grammar] || options[:grammar_url] || options[:limit] || options[:terminator] || raise(ArgumentError, "You must specify at least one of limit, terminator or grammar")
 
-      output_document = output_formatter.ssml_for_collection(prompts)
       grammars = AskGrammarBuilder.new(options).grammars
+
+      output_document = prompts.empty? ? nil : output_formatter.ssml_for_collection(prompts)
 
       PromptBuilder.new(output_document, grammars, options).execute self
     end