+++ lib/adhearsion-asr/result.rb
@@ -12,6 +12,15 @@ module AdhearsionASR
       self[:utterance] = mode == :dtmf ? parse_dtmf(other) : other
     end
 
+    def response
+      Adhearsion.deprecated "#utterance"
+      utterance
+    end
+
+    def match?
+      status == :match
+    end
+
     private
 
     def parse_dtmf(dtmf)