+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # [develop](https://github.com/adhearsion/adhearsion-asr)
 
+# [v1.1.0](https://github.com/adhearsion/adhearsion-asr/compare/1.0.1...1.1.0) - [2014-01-02](https://rubygems.org/gems/adhearsion-asr/versions/1.1.0)
+  * Feature: Alias `Result#response` to `#utterance`, with a deprecation warning
+  * Feature: Add `#match?` predicate to `Result` object
+  * Feature: When no prompts are supplied, only execute an input component
+
 # [v1.0.1](https://github.com/adhearsion/adhearsion-asr/compare/1.0.0...1.0.1) - [2013-09-30](https://rubygems.org/gems/adhearsion-asr/versions/1.0.1)
   * Bugfix: A menu definition's block context is now available